Refactor client to use new API
This API client is new and unofficial so it may be more brittle, however
it does appear to provide predictions that the old API does not.

This adds a new dependency on `requests` to make my life simpler, and it
also gets rid of tests. There is no pre/post processing of data, so
there's really nothing to unit test.